+++ adcs_math/quaternion.h
@@ -12,7 +12,7 @@
 
 
 typedef struct Quaternion {
-	float scalar;
+	double scalar;
 	vec3  vector;
 } quat; //Use "quat" as the type.
 
@@ -25,7 +25,7 @@ typedef struct Quaternion {
  *
  * @return Void.
  */
-void quat_set(float scalar, vec3 vector, quat *output);
+void quat_set(double scalar, vec3 vector, quat *output);
 
 
 /**@brief Multiply a quaternion by a scalar value.
@@ -38,7 +38,7 @@ void quat_set(float scalar, vec3 vector, quat *output);
  *
  * @return Void.
  */
-void quat_scalar(float scalar, quat quaternion, quat *output);
+void quat_scalar(double scalar, quat quaternion, quat *output);
 
 
 /**@brief Multiply two quaternions.
@@ -96,13 +96,13 @@ void quat_inverse(quat quaternion, quat *output);
 
 /**@brief Create a quaternion from an axis and rotation angle.
  *
- * @param angle The angle of rotation around the axis of rotation.
+ * @param angle The angle of rotation in radians around the axis.
  * @param vector The vector whose direction is the axis of rotation. 
  * @param output The quat* that will hold the resulting quaternion.
  *
  * @return Void.
  */
-void quat_from(float angle, vec3 vector, quat *output);
+void quat_from(double angle, vec3 vector, quat *output);
 
 
 /**@brief Apply the rotation represented by a quaternion to a vector.
@@ -122,7 +122,7 @@ void quat_rotate_vec(vec3 vector, quat quaternion, vec3 *output);
  *
  * @return The magnitude of the quaternion.
  */
-float quat_mag(quat quaternion);
+double quat_mag(quat quaternion);
 
 
 #endif//QUATERNION_H
